# Speedup `Tr[A @ B]` by computing `(A * B).sum()`