--- a/docs/python_api/index.md
+++ b/docs/python_api/index.md
@@ -159,6 +159,7 @@ The **dynamiqs** Python API features two main types of functions: solvers of dif
         members:
         - dag
         - mpow
+        - tracemm
         - trace
         - ptrace
         - tensor