Thanks for the tip @gaspardbb!

- I added this to the public API bc I think it's a fairly common operation.
- Fixed `overlap()` that had an extra incorrect `squeeze()` for density matrices.
- Used the new `tracemm()` in `expect()` and `overlap()`.

Benchmark for small matrices is already impressive (jitted on CPU):
```text
=== n=16
5.97 µs ± 159 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
5.31 µs ± 506 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
=== n=128
190 µs ± 2.99 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
19.9 µs ± 836 ns per loop (mean ± std. dev. of 7 runs, 100 loops each)
=== n=256
935 µs ± 85 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
96.1 µs ± 4.21 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

Here's how the doc looks like:
![Screenshot 2024-02-28 at 20 24 34](https://github.com/dynamiqs/dynamiqs/assets/38159029/57c3db15-46b1-4b28-9c77-fd5cb5a51673)